+++ src/movements/PathMovement.cpp
@@ -441,10 +441,10 @@ void PathMovement::set_snapping_trajectory(const Point& src, const Point& dst) {
 }
 
 /**
- * @brief Returns the current angle with the horizontal in radian
- * @return a angle in radian
+ * @brief Returns the current angle with the horizontal in radians.
+ * @return The angle in radians.
  */
-float PathMovement::get_angle() const {
+double PathMovement::get_angle() const {
     return get_current_direction() * Geometry::PI_OVER_4;
 }
